% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_build_graph.R
\name{build_graph_MRCO}
\alias{build_graph_MRCO}
\title{Create Transition Matrix for MRCO}
\usage{
build_graph_MRCO(
  cm,
  cells_selected,
  metadata,
  metadata_column_name,
  metadata_column_nbins,
  silent,
  warnings,
  igraph_layout_type
)
}
\arguments{
\item{cm}{abbreviation for cluster matrix: data.frame, tibble or matrix excerpt from metadata, only containing each resolution step (in order)}

\item{cells_selected}{contains which cellnames are selected}

\item{metadata}{data.frame or tibble with dim-names: row names cell identities and column names cell level metadata variables}

\item{metadata_column_name}{character or tidy-selection style unquoted name of a metadata column to plot piechart nodes from}

\item{metadata_column_nbins}{NULL, "all" or a numeric, set the number of bins to create from given metadata column if it is continuous data}

\item{silent}{logical, FALSE to suppress messages and warnings}

\item{warnings}{logical, FALSE to supress warnings}

\item{igraph_layout_type}{character giving the igraph layout type for graph creation; either "tree" or "sugiyama"}
}
\description{
Create Transition Matrix for MRCO
}
