% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cluster_stability.R
\name{calculate_cluster_scatter}
\alias{calculate_cluster_scatter}
\title{Cluster Scatter Calculator Optimized}
\usage{
calculate_cluster_scatter(
  count_matrix = count_matrix,
  cm = cm,
  graph_layout = graph_layout,
  silent = silent
)
}
\arguments{
\item{count_matrix}{sparse/matrix, containing raw/normalized or transformed gene expressions of each cell}

\item{cm}{cluster matrix, exerpt of metadata of seruat object, containing each cluster nr per resolution for each cell}

\item{graph_layout}{tibble, used graph_layout to plot and to be summarized}
}
\description{
Cluster Scatter Calculator Optimized
}
