% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_stability.R
\name{cluster_stability_estimator_MRCO}
\alias{cluster_stability_estimator_MRCO}
\title{Cluster Stability Estimator}
\usage{
cluster_stability_estimator_MRCO(
  graph_layout,
  edges,
  branch_df,
  edge_ratio_weigth = 0.9,
  plot = T
)
}
\arguments{
\item{graph_layout}{contains the to be plotted graph as tibble}

\item{edges}{tibble containing from and to of each edge}

\item{branch_df}{matrix, each column is one branch from root to leaf, length of list is nr of leafs}

\item{edge_ratio_weigth}{numeric between 0 and 1, when stable edges are determined their edge ratio must be larger than the branch paths maximum edge ratio timed edge_ratio_weight.
Therefore, a value closer to 1 is less permissive towards noise, whereas a value closer to 0 may handle noisier graphs better.}

\item{plot}{logical, FALSE to skip plotting}
}
\description{
Estimate stable cluster twig in relation to whole graph by four measures:
\itemize{
 \item{"Mean edge proportional size of twig"}{\cr UPDATE DEBUG A twig, clusters of incremental resolution on same x axis position, may be declared to be stable if the edge proportion (edge_size/from_node_size) is always close to 1. Therefore the mean of that edge sequence will also be close to 1.}
 \item{"Number of clusters within a twig"}{\cr UPDATE DEBUG A twig is likely to be stable if it contains a large proportion of available clusters. Available clusters will be the same across each branch of the graph, one per resolution that is tested.}
}
Due to many possible options the best recombination is still being tested, for now the following parameters are used to estimate best cluster choices:
\itemize{
 \item {"Mean edge proportional size of twig"}{\cr UPDATE DEBUG edge_size/from_node_size where the first element of each twig is skipped. This reduces dependence of twigs creation eg if the first cluster of the twig derives from a perfect 50/50 split then from_cluster_size will be twice as large leaving the proportional edge size crippled though the rest of the twig might be perfect cell conservation.}
 \item {"Number of clusters within a twig"}{\cr UPDATE DEBUG n_cluster/resolution_steps where the (vertical) nr of clusters available is equal to the resolution steps. We take the proportion to avoid biases across different graph strucutres.}
}
}
