% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_branching.R
\name{create_node_branch_list_bf_MRCO}
\alias{create_node_branch_list_bf_MRCO}
\title{Create node branch list breath first function}
\usage{
create_node_branch_list_bf_MRCO(
  graph_layout = graph_layout,
  edges = edges,
  edge_num_size_filter = 0,
  edge_prop_size_filter = 0.05,
  reduce = T,
  debugg = T,
  silent = F,
  branches_overflow = 5000
)
}
\arguments{
\item{graph_layout}{active graph layout that is used, at this stage with unique x positions per unique branch}

\item{edges}{edge list}

\item{edge_num_size_filter}{numeric, give the number of cells that must be surpassed to draw edge, has to be set for each data set as its highly dependent on nr of cells within data}

\item{edge_prop_size_filter}{numeric, give the proportion of cells that must move between nodes to draw edge}

\item{reduce}{logical, TRUE to walk as few paths as possible while still visiting every node, FALSE to walk any possible path passing the edge_filters; warning memory expensive, adjust branches_overflow parameter! Only relevant if suggest_cut is TRUE.}

\item{debugg}{logical, if TRUE prints mesage regarding skipping branches, default to false}

\item{silent}{logical, FALSE to suppress messages and warnings}

\item{branches_overflow}{numeric giving the max number of paths through the graph that is checked for each resolution step to protect memory overflow}
}
\description{
Create node branch list breath first function
}
