% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_branching.R
\name{create_node_branch_list_MRCO}
\alias{create_node_branch_list_MRCO}
\title{Create node branch list breath first function}
\usage{
create_node_branch_list_MRCO(graph_layout = graph_layout, edges = edges)
}
\arguments{
\item{graph_layout}{active graph layout that is used, at this stage with unique x positions per unique branch}

\item{edges}{edge list}
}
\description{
Create node branch list breath first function
}
