% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGCS.R
\name{create_transtion_matrix}
\alias{create_transtion_matrix}
\title{Create Transition Matrix wNGCS}
\usage{
create_transtion_matrix(cm = cm, nodes = nodes)
}
\arguments{
\item{cm}{synonym to cm, cluster matrix: data.frame, tibble or matrix excerpt from metadata, only containing each resolution step (in order)}

\item{nodes}{tibble of unique nodes, meaning for each resolution each cluster}
}
\description{
Create Transition Matrix wNGCS
}
