% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_build_graph.R
\name{create_transtion_matrix_MRCO}
\alias{create_transtion_matrix_MRCO}
\title{Create Transition Matrix MRCO}
\usage{
create_transtion_matrix_MRCO(cm = cm, nodes = nodes)
}
\arguments{
\item{cm}{synonym to cm, cluster matrix: data.frame, tibble or matrix
excerpt from metadata, only containing each resolution step (in order)}

\item{nodes}{tibble of unique nodes, meaning for each resolution each cluster}
}
\value{
data.frame tibble of edges from given nodes
}
\description{
Create Transition Matrix MRCO
}
