% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plot_graph.R
\name{createplot_labels}
\alias{createplot_labels}
\title{create plot labels MRCO}
\usage{
createplot_labels(plot_MRCO = NULL, highlight_selection = NULL)
}
\arguments{
\item{plot_MRCO}{ggplot, MRCO plot without labels}

\item{highlight_selection}{logical, TRUE to highlight selected nodes}
}
\value{
ggplot MRCO plot with labels
}
\description{
create MRCO plot labels
}
