% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plot_graph.R
\name{createplot_piechart_nodes}
\alias{createplot_piechart_nodes}
\title{create piechart nodes MRCO}
\usage{
createplot_piechart_nodes(
  plot_MRCO,
  metadata_column_name,
  graph_arc,
  graph_layout,
  highlight_selection
)
}
\arguments{
\item{plot_MRCO}{ggplot, MRCO plot without nodes}

\item{metadata_column_name}{character or tidy-selection style unquoted name of a metadata column to plot piechart nodes from}

\item{graph_arc}{tibble, used graph_arc to plot piecharts}

\item{graph_layout}{tibble, used graph_layout to plot and to be summarized}

\item{highlight_selection}{logical, TRUE to highlight selected nodes}
}
\value{
ggplot MRCO plot with piechart nodes
}
\description{
add nodes to existing edge graph for piechart plot
}
