% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plot_graph.R
\name{createplot_point_nodes}
\alias{createplot_point_nodes}
\title{create simple nodes MRCO}
\usage{
createplot_point_nodes(plot_MRCO, graph_layout, highlight_selection)
}
\arguments{
\item{plot_MRCO}{ggplot, MRCO plot with only edges}

\item{graph_layout}{tibble, used graph_layout to plot and to be summarized}

\item{highlight_selection}{logical, TRUE to highlight selected nodes}
}
\value{
ggplot MRCO plot with point nodes
}
\description{
add nodes to existing edge graph for simplenode plot
}
