% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plot_graph.R
\name{createplot_simplenode}
\alias{createplot_simplenode}
\title{create simple node plot MRCO}
\usage{
createplot_simplenode(
  plot_MRCO,
  graph_layout,
  highlight_selection = TRUE,
  no_labels
)
}
\arguments{
\item{plot_MRCO}{ggplot object, used to gradually build up final MRCO plot}

\item{graph_layout}{tibble, used graph_layout to plot and to be summarized}

\item{highlight_selection}{logical, TRUE to highlight selected nodes}

\item{no_labels}{logical, TRUE to turn off cluster labels, FALSE default for labeled clusters}
}
\value{
ggplot simplenode plot
}
\description{
create MRCO plot with simple nodes (single colour per node).
}
