% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGCS.R
\name{custom_node_selection}
\alias{custom_node_selection}
\title{Custom Node  Selection wNGCS}
\usage{
custom_node_selection(
  nodes_cell,
  nodes_selection,
  cm,
  merge_downwards,
  silent = T
)
}
\arguments{
\item{nodes_cell}{tibble of unique nodes, meaning for each resolution each cluster}

\item{nodes_selection}{vector or list, two ways to manually select a node by its global id from the graph:
either give resolution step and cluster id from that resolution as underscore separated character string,
 or give a list named where each name corresponds to a resolution step and its element is a single vector containing the cluster ids that are to be selected of that resolution.}

\item{cm}{cluster matrix to match outputs row order}

\item{merge_downwards}{logical, cells which are in multiple selected clusters can either be assorted into the first or last cluster available for them.
FALSE means that clusters are merged from the bottom towards the top of the graph, hence cells which are part of a higher resolution cluster will remain in it.
TRUE means that clusters are merged from top towards the bottom, hence cells will end up in the lowest resolution cluster that they may end in.}

\item{silent}{logical, FALSE to suppress messages and warnings}
}
\description{
Custom Node  Selection wNGCS
}
