% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_input_checker.R
\name{input_check_clustermatrix_MRCO}
\alias{input_check_clustermatrix_MRCO}
\title{Input checker cluster matrix MRCO}
\usage{
input_check_clustermatrix_MRCO(metadata = NULL, clustering_columns = NULL)
}
\arguments{
\item{metadata}{data.frame or tibble with dim-names: row names cell identities and column names cell level metadata variables}

\item{clustering_columns}{character or tidyselection call}
}
\value{
data.frame of all clustering resolutions selected from metadata
}
\description{
Input checker cluster matrix MRCO
}
