% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_input_checker.R
\name{input_check_metadata_MRCO}
\alias{input_check_metadata_MRCO}
\title{Input checker metadata MRCO}
\usage{
input_check_metadata_MRCO(
  metadata = NULL,
  metadata_column_name = NULL,
  nbins = NULL
)
}
\arguments{
\item{metadata}{data.frame or tibble: with row names (cell ids) and
column names (cell metadata variables)}

\item{metadata_column_name}{character or unquoted name: name the metadata
column to plot piechart nodes from}

\item{nbins}{numeric, set the number of bins to create from given metadata
column if it is numeric data}
}
\value{
list of updated input variables
}
\description{
Input checker metadata MRCO
}
