% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plot_graph.R
\name{plot_graph_MRCO}
\alias{plot_graph_MRCO}
\title{Plot Graph Structure MRCO}
\usage{
plot_graph_MRCO(
  metadata_column_name,
  graph_layout,
  graph_arc,
  edge_num_size_filter,
  edge_prop_size_filter,
  highlight_selection = TRUE,
  no_labels
)
}
\arguments{
\item{metadata_column_name}{character or tidy-selection style unquoted name of a metadata column to plot piechart nodes from}

\item{graph_layout}{tibble, used graph_layout to plot and to be summarized}

\item{graph_arc}{tibble, used graph_arc to plot piecharts}

\item{edge_num_size_filter}{numeric, give the number of cells that must be surpassed to draw edge, has to be set for each data set as its highly dependent on nr of cells within data}

\item{edge_prop_size_filter}{numeric, give the proportion of cells that must move between nodes to draw edge}

\item{highlight_selection}{logical, TRUE to highlight selected nodes}

\item{no_labels}{logical, TRUE to turn off cluster labels, FALSE default for labeled clusters}
}
\value{
ggplot MRCO final plot
}
\description{
uses MRCO internal arguments to plot tree like layouts
}
