% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGCS_plot_graph.R
\name{plot_graph_NGCS}
\alias{plot_graph_NGCS}
\title{Plot Graph Structure NGCS}
\usage{
plot_graph_NGCS(
  metadata_column_name,
  graph_layout,
  graph_arc,
  edge_num_size_filter,
  edge_prop_size_filter,
  highlight_selection = T,
  plot_col_gradient,
  no_labels
)
}
\arguments{
\item{metadata_column_name}{character or tidy-selection style unquoted name of a metadata column to plot piechart nodes from}

\item{graph_layout}{tibble, used graph_layout to plot and to be summarized}

\item{graph_arc}{tibble, used graph_arc to plot piecharts}

\item{edge_num_size_filter}{numeric, give the number of cells that must be surpassed to draw edge, has to be set for each data set as its highly dependent on nr of cells within data}

\item{edge_prop_size_filter}{numeric, give the proportion of cells that must move between nodes to draw edge}

\item{highlight_selection}{logical, TRUE to highlight selected nodes}

\item{plot_col_gradient}{NULL for auto detect, or logical, TRUE to use continuous fill, FALSE for discrete fill; only applies when continuous metadata column is selected}

\item{no_labels}{logical, TRUE to turn off cluster labels, FALSE default for labeled clusters}
}
\description{
uses NGCS internal arguments to plot tree like layouts
}
