% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_build_graph.R
\name{check_unique_x_positions_MRCO}
\alias{check_unique_x_positions_MRCO}
\title{check for uniqueness of x nodes in MRCO graph}
\usage{
check_unique_x_positions_MRCO(graph_layout = graph_layout, silent = silent)
}
\arguments{
\item{graph_layout}{tibble, used graph_layout to plot and to be summarized}

\item{silent}{logical, FALSE to suppress messages and warnings}
}
\description{
check for uniqueness of x nodes in MRCO graph
}
